Improve the performance of Display.
Or more accurately reduce the number of times write_str is called in the
common case. Prior to this change any partial chunks at the end of input
would be written in a separate write_str call. Since write_str is very
commonly the slowest part this proved costly. Instead we know append the
encoded partial chunk to the output buffer when there is enough space
remaining.