Fix a soundness hole in the block iterator.
The block iterator provides chunks that potentially overlap. Due to this
it's ill-advised to use the standard iterator trait because it's
possible to have chunks from different iterations live at the same time
and that would violate the aliasing rules (there would be two mutable
references to the same memory). This is just an internal implementation
and the aliasing rules can not be violated by anything other than the
internals of the crate, but it's still better to avoid the possibility
of misuse even internally.

The fix is pretty straightforward. Instead of implementing the Iterator
interface instead implement an inhernent method called next_chunk that
returns values tied to the lifetime of the self borrow. Effectively a
"streaming iterator". The main drawback is that we now need to use
`while let` loops instead of `for` loops, but otherwise a pretty
straightforward change.